Rename package ai.vespa.hosted.cd.metrics to .metric
