+++ tenant-cd/src/main/java/ai/vespa/hosted/cd/metric/Space.java
@@ -1,4 +1,4 @@
-package ai.vespa.hosted.cd.metrics;
+package ai.vespa.hosted.cd.metric;
 
 import java.util.List;
 import java.util.Map;
