+++ tenant-cd/src/main/java/ai/vespa/hosted/cd/metric/Statistic.java
@@ -1,4 +1,4 @@
-package ai.vespa.hosted.cd.metrics;
+package ai.vespa.hosted.cd.metric;
 
 import java.util.HashMap;
 import java.util.Map;
